SDL: Move the Capabilities struct into OpenGLSdlGraphicsManager
